Print definition of received error
This does not match the long description of the error but shows it so
that it becomes easier to debug. Note that two different instances of
the same error type would still match.

This is unavoidable unless the match was changed so that the message is
also expected (and explicitly written) in the test cases.