Add debug config for Python with param prompt
This was added to allow to more comfortably launch the project in debug
mode with ad-hoc parameters that you can enter in the prompt. This is to
avoid debugging the entire test suite each time you try to debug a
specific case. You can simply filter using the parameters. Remember to
run this with the ./brewtest.py file active.